Use probed Qt directories on non-Windows platforms
Patch ea45279 added an assumption that the Qt headers are installed in
include/ subdirectory of a project root. However, if Qt4 is provided by
system packages on Linux, the headers and libraries will no longer be
sibling directories (i.e. it is /usr/include/qt4 instead of
/usr/local/qt4/include).

This patch restores the old functionality of using the probed
variables. (I haven't tested whether these really should be used for
Windows also; I think that should be considered).