Remark That 3pt EPS Changes Function Shape in [Sr, Smax]
The three-point vertical scaling option will usually change the
shape of the relative permeability function in the saturation
interval between the scaled critical displacing saturation and the
maximum phase saturation.  In particular, we usually replace a
general, non-linear function with a linear approximation when
applying three-point vertical scaling.

Add a comment to that effect.