Don't Calculate Scaled Curves if EPS Data Unavailable
This commit captures the result set's availability of EPS data
in a new data member,

    bool ECLSaturationFunc::Impl::haveEPSData_

We need this piece of information to identify whether a null
'eps_' data member corresponds to missing EPS data or failure to
form the backing EPS data object.  In the first case, it is fine
to output the unscaled curves while throwing an exception in the
latter case is more appropriate.

This restores ResInsight's ability to show unscaled saturation
function curves if the backing EPS data is not output to the
result set's INIT file.