#5696 WBS: Fix check for element properties when creating options source list.
An off-by-one error in the time step when checking if a element property table
has any values would lead to the "Propery Table" option not being shown in the
UI for the default time step (i.e 0).

This makes it possible to select UCS and POISSON_RATIO as element properties.