#7774 Ensemble Statistics : Remove test for match on first reported simulation time step
When computing statistics curves for an ensemble, we computed these values based on curves having identical same timestamp as the first summary case in the ensemble. If the simulator had numerical issues, this first time step can be slightly different from case to case.

Remove this test on first time step, but keep the optional test on last time step.