AutoVcpkg: Copy bootstrap file from cmake folder
AutoVcpkg assumed that the bootstrap file was in the root project
folder, and not in a cmake subfolder.

Note: It is not correct to use ${CMAKE_CURRENT_LIST_DIR} since this is
inside a function and whould still refer to the folder of the calling
file. The more elegant solution would be to use
${CMAKE_CURRENT_FUNCTION_LIST_DIR}, but that is not available until
CMake 3.17.