AutoVcpkg: Include Vcpkg toolchain
Include the Vcpkg toolchain file after the installation of the first
package.

When building for the first time from a clean source tree, the Vcpkg
toolchain file does not exist. At the same time CMake is only reading
the toolchain file the very first time it is configuring, but since the
toolchain file is only created when vcpkg is configured on the first
call to vcpkg_install it will never read it. So it is not enough to just
reconfigure.

Note: It is not ideal to include it here since one might at some point
install and search for a package at an erlier point in the script and
fail to find it.

Alternative approaces that I considered:
- Include this file right after including AutoVcpkg. This would require
  1 recompilation for the file to be found and read which is
  unfortunate.
- Include this file in the config function in AutoVcpkg. This seemed
  difficult without including the file for every call to vcpkg_install,
  which seems unecessary.