Fix intersection info for thin and twisted cells
A cell face is split into four triangles using the midpoint of the cell. The algorithm for intersecting a line segment with triangles might intersect the geometry of the cell face at the opposite cell face before the first cell face.
 
This behaviour has been seen for thin layers cells with heavy geometry twisting.

The fix will be to switch the isEnteringCell flag for both intersection.