Refactor PdmChildArrayField to use more descriptive function names
This commit changes the function names in PdmChildArrayField to be more descriptive. Specifically, children() is renamed to childrenByType(), and children() that takes a vector pointer argument is replaced with a new function called children() that returns a vector of child objects. Additionally, some functions are marked as virtual and given empty implementations. Finally, hasChildren() is updated to use the new children() implementation.