Summary Plot: Compute default y-range based on visible plots
Previously, when the stepping dimension was set to 'well' for range aggregation, it was based on all wells. If one of the wells had extreme values and was not visible, it would set the y-range to match the extreme value, making some curves invisible. We have now changed the default setting to use visible subplots to determine the y-range aggregation.