#11393 Contour map: Make sure polygon cell filter works as expected
Add the polygonInViewCollection to display the polygon in tree view. This makes the polygon filter work, and interaction in contour map is also supported.

Polygon cell filters are not working due to different organization of the project tree for Eclipse view and contour map view. Manually assign pointers to polygon objects.