#11626 Python: By default, use port number assigned by GRPC
Using ResInsight in a server environment when many multiple versions of ResInsight is launched, can cause the port assignment to fail. Change the default value for launch_port from -1 to 0, as this will improve the logic used to assign a port to a new ResInsight session.

Using 0 will ask GRPC to find an available port to be used.