Bump minimum required CMake to 3.26
* Bump minimum required CMake to 3.26

Recently, dependencies using vcpkg was updated, including boost 1.86. When boost 1.86 is linked by ResInsight, the minimum required CMake version is 3.26.

It is possible to manage dependencies manually without using vcpkg. In this case, the ResInsight build might work with a lower CMake requirement. This scenario is not tested by the ResInsight team.

* Use minimum required version of CMake when building on GitHub Action