#12262 Make sure signals are disconnected before objects are deleted
We can have large collections of object, and the delete of these objects can be time consuming. To improve the user experience, the delete of these objects is delegated to `PdmChildArrayField<DataType*>::deleteChildrenAsync()`. When an object is deleted, it will disconnect signals from observers.

When the owner object is deleted all connected signals is disconnected. If the owner object is deleted at the same time as the contained objects are deleted in a separate thread we will quite often have undefined behaviour. This undefined behaviour is reproducible in a unit test.

Make sure that all signals are disconnected before the delete operation.