Install ResInsight Application on macOS
This adds support to install the Application folder when the OS is macOS. In particular, the bundle is being created in the root build directory and installed in the root path of the installation prefix. Additionally, install the ERT libraries in the Contents/Frameworks bundle folder where the RPATH of the is automatically set to find them. Otherwise a link error arises.