Refactors PDM field reading for migration
Updates field reading to allow for data migration during the XML parsing process.

This change introduces a `migrateFieldContent` virtual method to `PdmObjectHandle`, allowing derived classes to modify field content before it's applied to the field.
The `PdmFieldReader` is refactored to leverage this method when available.
The specialized QString reader is merged into the generic reader.