#12953 Always check date of all realizations
The date stamp of UNSMRY is compared to date stamp of ESMRY. This takes some time, and a performance improvement was seen when this operation was done only on a single representative realization of an ensemble. This is not stable, as a single realization might crash or have no new data compared to other realizations.

This feature is removed.